@@ -2,13 +2,13 @@
 name = "feast"
 description = "Python SDK for Feast"
 readme = "README.md"
-requires-python = ">=3.9.0"
+requires-python = ">=3.10.0"
 license = {file = "LICENSE"}
 classifiers = [
     "License :: OSI Approved :: Apache Software License",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9"
+    "Programming Language :: Python :: 3.10"
 ]
 dynamic = [
     "version",
@@ -61,7 +61,7 @@ delta = ["deltalake"]
 docling = ["docling==2.27.0"]
 duckdb = ["ibis-framework[duckdb]>=9.0.0,<=9.5.0"]
 elasticsearch = ["elasticsearch>=8.13.0"]
-faiss = ["faiss-cpu>=1.7.0,<2"]
+faiss = ["faiss-cpu>=1.7.0,<=1.10.0"]
 gcp = [
     "google-api-core>=1.23.0,<3",
     "googleapis-common-protos>=1.52.0,<2",
@@ -75,9 +75,9 @@ gcp = [
 ge = ["great_expectations>=0.15.41,<1"]
 go = ["cffi>=1.15.0"]
 grpcio = [
-    "grpcio>=1.56.2,<2",
-    "grpcio-reflection>=1.56.2,<2",
-    "grpcio-health-checking>=1.56.2,<2",
+    "grpcio>=1.56.2,<=1.62.3",
+    "grpcio-reflection>=1.56.2,<=1.62.3",
+    "grpcio-health-checking>=1.56.2,<=1.62.3",
 ]
 hazelcast = ["hazelcast-python-client>=5.1"]
 hbase = ["happybase>=1.2.0,<3"]
@@ -123,8 +123,8 @@ ci = [
     "cryptography>=43.0,<44",
     "ruff>=0.8.0",
     "mypy-protobuf>=3.1",
-    "grpcio-tools>=1.56.2,<2",
-    "grpcio-testing>=1.56.2,<2",
+    "grpcio-tools>=1.56.2,<=1.62.3",
+    "grpcio-testing>=1.56.2,<=1.62.3",
     # FastAPI does not correctly pull starlette dependency on httpx see thread(https://github.com/tiangolo/fastapi/issues/5656).
     "httpx==0.27.2",
     "minio==7.2.11",
@@ -227,7 +227,7 @@ tag_regex = "^(?:[\\/\\w-]+)?(?P<version>[vV]?\\d+(?:\\.\\d+){0,2}[^\\+]*)(?:\\+
 
 [tool.ruff]
 line-length = 88
-target-version = "py39"
+target-version = "py310"
 include = ["*.py", "*.pyi"]
 
 [tool.ruff.format]
