@@ -17,11 +17,16 @@ limitations under the License.
 package e2e
 
 import (
+	"fmt"
+	"os"
+
 	"github.com/feast-dev/feast/infra/feast-operator/test/utils"
 	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
 )
 
 var _ = Describe("controller", Ordered, func() {
+	_, isRunOnOpenShiftCI := os.LookupEnv("RUN_ON_OPENSHIFT_CI")
 	featureStoreName := "simple-feast-setup"
 	feastResourceName := utils.FeastPrefix + featureStoreName
 	feastK8sResourceNames := []string{
@@ -29,26 +34,43 @@ var _ = Describe("controller", Ordered, func() {
 		feastResourceName + "-offline",
 		feastResourceName + "-ui",
 	}
+	namespace := "test-ns-feast"
+	defaultFeatureStoreCRTest := "TesDefaultFeastCR"
+	remoteRegistryFeatureStoreCRTest := "TestRemoteRegistryFeastCR"
+	applyAndMaterializeTest := "TestApplyAndMaterializeFeastDefinitions"
 
 	runTestDeploySimpleCRFunc := utils.GetTestDeploySimpleCRFunc("/test/e2e",
 		"test/testdata/feast_integration_test_crs/v1alpha1_default_featurestore.yaml",
-		featureStoreName, feastResourceName, feastK8sResourceNames)
+		featureStoreName, feastResourceName, feastK8sResourceNames, namespace)
 
 	runTestWithRemoteRegistryFunction := utils.GetTestWithRemoteRegistryFunc("/test/e2e",
 		"test/testdata/feast_integration_test_crs/v1alpha1_default_featurestore.yaml",
 		"test/testdata/feast_integration_test_crs/v1alpha1_remote_registry_featurestore.yaml",
-		featureStoreName, feastResourceName, feastK8sResourceNames)
+		featureStoreName, feastResourceName, feastK8sResourceNames, namespace)
+
+	runTestApplyAndMaterializeFunc := utils.RunTestApplyAndMaterializeFunc("/test/e2e", namespace, "credit-scoring", utils.FeastPrefix+"credit-scoring")
 
 	BeforeAll(func() {
-		utils.DeployOperatorFromCode("/test/e2e", false)
+		if !isRunOnOpenShiftCI {
+			utils.DeployOperatorFromCode("/test/e2e", false)
+		}
+		By(fmt.Sprintf("Creating test namespace: %s", namespace))
+		err := utils.CreateNamespace(namespace, "/test/e2e")
+		Expect(err).ToNot(HaveOccurred(), fmt.Sprintf("failed to create namespace %s", namespace))
 	})
 
 	AfterAll(func() {
-		utils.DeleteOperatorDeployment("/test/e2e")
+		if !isRunOnOpenShiftCI {
+			utils.DeleteOperatorDeployment("/test/e2e")
+		}
+		By(fmt.Sprintf("Deleting test namespace: %s", namespace))
+		err := utils.DeleteNamespace(namespace, "/test/e2e")
+		Expect(err).ToNot(HaveOccurred(), fmt.Sprintf("failed to delete namespace %s", namespace))
 	})
 
 	Context("Operator E2E Tests", func() {
-		It("Should be able to deploy and run a default feature store CR successfully", runTestDeploySimpleCRFunc)
-		It("Should be able to deploy and run a feature store with remote registry CR successfully", runTestWithRemoteRegistryFunction)
+		It("Should be able to deploy and run a "+defaultFeatureStoreCRTest+" successfully", runTestDeploySimpleCRFunc)
+		It("Should be able to deploy and run a "+remoteRegistryFeatureStoreCRTest+"  successfully", runTestWithRemoteRegistryFunction)
+		It("Should be able to apply and run a "+applyAndMaterializeTest+" materialize feature store definitions using CronJob successfully", runTestApplyAndMaterializeFunc)
 	})
 })
