@@ -35,8 +35,11 @@ FEATURE_TRANSFORMATION_SERVER_PORT_ENV_NAME: str = "FEATURE_TRANSFORMATION_SERVE
 # Default FTS port
 DEFAULT_FEATURE_TRANSFORMATION_SERVER_PORT = 6569
 
-# Default registry server port
+# Default registry server grpc port
 DEFAULT_REGISTRY_SERVER_PORT = 6570
+
+# Default registry server REST port
+DEFAULT_REGISTRY_REST_SERVER_PORT = 6572
 
 # Default offline server port
 DEFAULT_OFFLINE_SERVER_PORT = 8815
