@@ -1,22 +1,12 @@
 # -*- coding: utf-8 -*-
 # Generated by the protocol buffer compiler.  DO NOT EDIT!
-# NO CHECKED-IN PROTOBUF GENCODE
 # source: feast/core/StreamFeatureView.proto
-# Protobuf Python Version: 5.29.0
+# Protobuf Python Version: 4.25.1
 """Generated protocol buffer code."""
 from google.protobuf import descriptor as _descriptor
 from google.protobuf import descriptor_pool as _descriptor_pool
-from google.protobuf import runtime_version as _runtime_version
 from google.protobuf import symbol_database as _symbol_database
 from google.protobuf.internal import builder as _builder
-_runtime_version.ValidateProtobufRuntimeVersion(
-    _runtime_version.Domain.PUBLIC,
-    5,
-    29,
-    0,
-    '',
-    'feast/core/StreamFeatureView.proto'
-)
 # @@protoc_insertion_point(imports)
 
 _sym_db = _symbol_database.Default()
@@ -36,12 +26,12 @@ DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\"feast/core/Strea
 _globals = globals()
 _builder.BuildMessageAndEnumDescriptors(DESCRIPTOR, _globals)
 _builder.BuildTopDescriptorsAndMessages(DESCRIPTOR, 'feast.core.StreamFeatureView_pb2', _globals)
-if not _descriptor._USE_C_DESCRIPTORS:
-  _globals['DESCRIPTOR']._loaded_options = None
+if _descriptor._USE_C_DESCRIPTORS == False:
+  _globals['DESCRIPTOR']._options = None
   _globals['DESCRIPTOR']._serialized_options = b'\n\020feast.proto.coreB\026StreamFeatureViewProtoZ/github.com/feast-dev/feast/go/protos/feast/core'
-  _globals['_STREAMFEATUREVIEWSPEC_TAGSENTRY']._loaded_options = None
+  _globals['_STREAMFEATUREVIEWSPEC_TAGSENTRY']._options = None
   _globals['_STREAMFEATUREVIEWSPEC_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_STREAMFEATUREVIEWSPEC'].fields_by_name['user_defined_function']._loaded_options = None
+  _globals['_STREAMFEATUREVIEWSPEC'].fields_by_name['user_defined_function']._options = None
   _globals['_STREAMFEATUREVIEWSPEC'].fields_by_name['user_defined_function']._serialized_options = b'\030\001'
   _globals['_STREAMFEATUREVIEW']._serialized_start=268
   _globals['_STREAMFEATUREVIEW']._serialized_end=379
