@@ -1,22 +1,12 @@
 # -*- coding: utf-8 -*-
 # Generated by the protocol buffer compiler.  DO NOT EDIT!
-# NO CHECKED-IN PROTOBUF GENCODE
 # source: feast/registry/RegistryServer.proto
-# Protobuf Python Version: 5.29.0
+# Protobuf Python Version: 4.25.1
 """Generated protocol buffer code."""
 from google.protobuf import descriptor as _descriptor
 from google.protobuf import descriptor_pool as _descriptor_pool
-from google.protobuf import runtime_version as _runtime_version
 from google.protobuf import symbol_database as _symbol_database
 from google.protobuf.internal import builder as _builder
-_runtime_version.ValidateProtobufRuntimeVersion(
-    _runtime_version.Domain.PUBLIC,
-    5,
-    29,
-    0,
-    '',
-    'feast/registry/RegistryServer.proto'
-)
 # @@protoc_insertion_point(imports)
 
 _sym_db = _symbol_database.Default()
@@ -43,30 +33,30 @@ DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n#feast/registry/Re
 _globals = globals()
 _builder.BuildMessageAndEnumDescriptors(DESCRIPTOR, _globals)
 _builder.BuildTopDescriptorsAndMessages(DESCRIPTOR, 'feast.registry.RegistryServer_pb2', _globals)
-if not _descriptor._USE_C_DESCRIPTORS:
-  _globals['DESCRIPTOR']._loaded_options = None
+if _descriptor._USE_C_DESCRIPTORS == False:
+  _globals['DESCRIPTOR']._options = None
   _globals['DESCRIPTOR']._serialized_options = b'Z3github.com/feast-dev/feast/go/protos/feast/registry'
-  _globals['_LISTENTITIESREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTENTITIESREQUEST_TAGSENTRY']._options = None
   _globals['_LISTENTITIESREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTDATASOURCESREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTDATASOURCESREQUEST_TAGSENTRY']._options = None
   _globals['_LISTDATASOURCESREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTFEATUREVIEWSREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTFEATUREVIEWSREQUEST_TAGSENTRY']._options = None
   _globals['_LISTFEATUREVIEWSREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTALLFEATUREVIEWSREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTALLFEATUREVIEWSREQUEST_TAGSENTRY']._options = None
   _globals['_LISTALLFEATUREVIEWSREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTSTREAMFEATUREVIEWSREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTSTREAMFEATUREVIEWSREQUEST_TAGSENTRY']._options = None
   _globals['_LISTSTREAMFEATUREVIEWSREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTONDEMANDFEATUREVIEWSREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTONDEMANDFEATUREVIEWSREQUEST_TAGSENTRY']._options = None
   _globals['_LISTONDEMANDFEATUREVIEWSREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTFEATURESERVICESREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTFEATURESERVICESREQUEST_TAGSENTRY']._options = None
   _globals['_LISTFEATURESERVICESREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTSAVEDDATASETSREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTSAVEDDATASETSREQUEST_TAGSENTRY']._options = None
   _globals['_LISTSAVEDDATASETSREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTVALIDATIONREFERENCESREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTVALIDATIONREFERENCESREQUEST_TAGSENTRY']._options = None
   _globals['_LISTVALIDATIONREFERENCESREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTPERMISSIONSREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTPERMISSIONSREQUEST_TAGSENTRY']._options = None
   _globals['_LISTPERMISSIONSREQUEST_TAGSENTRY']._serialized_options = b'8\001'
-  _globals['_LISTPROJECTSREQUEST_TAGSENTRY']._loaded_options = None
+  _globals['_LISTPROJECTSREQUEST_TAGSENTRY']._options = None
   _globals['_LISTPROJECTSREQUEST_TAGSENTRY']._serialized_options = b'8\001'
   _globals['_REFRESHREQUEST']._serialized_start=487
   _globals['_REFRESHREQUEST']._serialized_end=520
