@@ -1,7 +1,6 @@
 # Generated by the gRPC Python protocol compiler plugin. DO NOT EDIT!
 """Client and server classes corresponding to protobuf-defined services."""
 import grpc
-import warnings
 
 from feast.protos.feast.core import DataSource_pb2 as feast_dot_core_dot_DataSource__pb2
 from feast.protos.feast.core import Entity_pb2 as feast_dot_core_dot_Entity__pb2
@@ -18,25 +17,6 @@ from feast.protos.feast.core import ValidationProfile_pb2 as feast_dot_core_dot_
 from feast.protos.feast.registry import RegistryServer_pb2 as feast_dot_registry_dot_RegistryServer__pb2
 from google.protobuf import empty_pb2 as google_dot_protobuf_dot_empty__pb2
 
-GRPC_GENERATED_VERSION = '1.71.0'
-GRPC_VERSION = grpc.__version__
-_version_not_supported = False
-
-try:
-    from grpc._utilities import first_version_is_lower
-    _version_not_supported = first_version_is_lower(GRPC_VERSION, GRPC_GENERATED_VERSION)
-except ImportError:
-    _version_not_supported = True
-
-if _version_not_supported:
-    raise RuntimeError(
-        f'The grpc package installed is at version {GRPC_VERSION},'
-        + f' but the generated code in feast/registry/RegistryServer_pb2_grpc.py depends on'
-        + f' grpcio>={GRPC_GENERATED_VERSION}.'
-        + f' Please upgrade your grpc module to grpcio>={GRPC_GENERATED_VERSION}'
-        + f' or downgrade your generated code using grpcio-tools<={GRPC_VERSION}.'
-    )
-
 
 class RegistryServerStub(object):
     """Missing associated documentation comment in .proto file."""
@@ -51,227 +31,227 @@ class RegistryServerStub(object):
                 '/feast.registry.RegistryServer/ApplyEntity',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplyEntityRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetEntity = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetEntity',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetEntityRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_Entity__pb2.Entity.FromString,
-                _registered_method=True)
+                )
         self.ListEntities = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListEntities',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListEntitiesRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListEntitiesResponse.FromString,
-                _registered_method=True)
+                )
         self.DeleteEntity = channel.unary_unary(
                 '/feast.registry.RegistryServer/DeleteEntity',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.DeleteEntityRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.ApplyDataSource = channel.unary_unary(
                 '/feast.registry.RegistryServer/ApplyDataSource',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplyDataSourceRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetDataSource = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetDataSource',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetDataSourceRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_DataSource__pb2.DataSource.FromString,
-                _registered_method=True)
+                )
         self.ListDataSources = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListDataSources',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListDataSourcesRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListDataSourcesResponse.FromString,
-                _registered_method=True)
+                )
         self.DeleteDataSource = channel.unary_unary(
                 '/feast.registry.RegistryServer/DeleteDataSource',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.DeleteDataSourceRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.ApplyFeatureView = channel.unary_unary(
                 '/feast.registry.RegistryServer/ApplyFeatureView',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplyFeatureViewRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.DeleteFeatureView = channel.unary_unary(
                 '/feast.registry.RegistryServer/DeleteFeatureView',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.DeleteFeatureViewRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetAnyFeatureView = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetAnyFeatureView',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetAnyFeatureViewRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.GetAnyFeatureViewResponse.FromString,
-                _registered_method=True)
+                )
         self.ListAllFeatureViews = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListAllFeatureViews',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListAllFeatureViewsRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListAllFeatureViewsResponse.FromString,
-                _registered_method=True)
+                )
         self.GetFeatureView = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetFeatureView',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetFeatureViewRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_FeatureView__pb2.FeatureView.FromString,
-                _registered_method=True)
+                )
         self.ListFeatureViews = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListFeatureViews',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListFeatureViewsRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListFeatureViewsResponse.FromString,
-                _registered_method=True)
+                )
         self.GetStreamFeatureView = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetStreamFeatureView',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetStreamFeatureViewRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_StreamFeatureView__pb2.StreamFeatureView.FromString,
-                _registered_method=True)
+                )
         self.ListStreamFeatureViews = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListStreamFeatureViews',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListStreamFeatureViewsRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListStreamFeatureViewsResponse.FromString,
-                _registered_method=True)
+                )
         self.GetOnDemandFeatureView = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetOnDemandFeatureView',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetOnDemandFeatureViewRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_OnDemandFeatureView__pb2.OnDemandFeatureView.FromString,
-                _registered_method=True)
+                )
         self.ListOnDemandFeatureViews = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListOnDemandFeatureViews',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListOnDemandFeatureViewsRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListOnDemandFeatureViewsResponse.FromString,
-                _registered_method=True)
+                )
         self.ApplyFeatureService = channel.unary_unary(
                 '/feast.registry.RegistryServer/ApplyFeatureService',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplyFeatureServiceRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetFeatureService = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetFeatureService',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetFeatureServiceRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_FeatureService__pb2.FeatureService.FromString,
-                _registered_method=True)
+                )
         self.ListFeatureServices = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListFeatureServices',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListFeatureServicesRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListFeatureServicesResponse.FromString,
-                _registered_method=True)
+                )
         self.DeleteFeatureService = channel.unary_unary(
                 '/feast.registry.RegistryServer/DeleteFeatureService',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.DeleteFeatureServiceRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.ApplySavedDataset = channel.unary_unary(
                 '/feast.registry.RegistryServer/ApplySavedDataset',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplySavedDatasetRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetSavedDataset = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetSavedDataset',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetSavedDatasetRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_SavedDataset__pb2.SavedDataset.FromString,
-                _registered_method=True)
+                )
         self.ListSavedDatasets = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListSavedDatasets',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListSavedDatasetsRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListSavedDatasetsResponse.FromString,
-                _registered_method=True)
+                )
         self.DeleteSavedDataset = channel.unary_unary(
                 '/feast.registry.RegistryServer/DeleteSavedDataset',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.DeleteSavedDatasetRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.ApplyValidationReference = channel.unary_unary(
                 '/feast.registry.RegistryServer/ApplyValidationReference',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplyValidationReferenceRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetValidationReference = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetValidationReference',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetValidationReferenceRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_ValidationProfile__pb2.ValidationReference.FromString,
-                _registered_method=True)
+                )
         self.ListValidationReferences = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListValidationReferences',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListValidationReferencesRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListValidationReferencesResponse.FromString,
-                _registered_method=True)
+                )
         self.DeleteValidationReference = channel.unary_unary(
                 '/feast.registry.RegistryServer/DeleteValidationReference',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.DeleteValidationReferenceRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.ApplyPermission = channel.unary_unary(
                 '/feast.registry.RegistryServer/ApplyPermission',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplyPermissionRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetPermission = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetPermission',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetPermissionRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_Permission__pb2.Permission.FromString,
-                _registered_method=True)
+                )
         self.ListPermissions = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListPermissions',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListPermissionsRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListPermissionsResponse.FromString,
-                _registered_method=True)
+                )
         self.DeletePermission = channel.unary_unary(
                 '/feast.registry.RegistryServer/DeletePermission',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.DeletePermissionRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.ApplyProject = channel.unary_unary(
                 '/feast.registry.RegistryServer/ApplyProject',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplyProjectRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetProject = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetProject',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetProjectRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_Project__pb2.Project.FromString,
-                _registered_method=True)
+                )
         self.ListProjects = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListProjects',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListProjectsRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListProjectsResponse.FromString,
-                _registered_method=True)
+                )
         self.DeleteProject = channel.unary_unary(
                 '/feast.registry.RegistryServer/DeleteProject',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.DeleteProjectRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.ApplyMaterialization = channel.unary_unary(
                 '/feast.registry.RegistryServer/ApplyMaterialization',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ApplyMaterializationRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.ListProjectMetadata = channel.unary_unary(
                 '/feast.registry.RegistryServer/ListProjectMetadata',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.ListProjectMetadataRequest.SerializeToString,
                 response_deserializer=feast_dot_registry_dot_RegistryServer__pb2.ListProjectMetadataResponse.FromString,
-                _registered_method=True)
+                )
         self.UpdateInfra = channel.unary_unary(
                 '/feast.registry.RegistryServer/UpdateInfra',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.UpdateInfraRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.GetInfra = channel.unary_unary(
                 '/feast.registry.RegistryServer/GetInfra',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.GetInfraRequest.SerializeToString,
                 response_deserializer=feast_dot_core_dot_InfraObject__pb2.Infra.FromString,
-                _registered_method=True)
+                )
         self.Commit = channel.unary_unary(
                 '/feast.registry.RegistryServer/Commit',
                 request_serializer=google_dot_protobuf_dot_empty__pb2.Empty.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.Refresh = channel.unary_unary(
                 '/feast.registry.RegistryServer/Refresh',
                 request_serializer=feast_dot_registry_dot_RegistryServer__pb2.RefreshRequest.SerializeToString,
                 response_deserializer=google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-                _registered_method=True)
+                )
         self.Proto = channel.unary_unary(
                 '/feast.registry.RegistryServer/Proto',
                 request_serializer=google_dot_protobuf_dot_empty__pb2.Empty.SerializeToString,
                 response_deserializer=feast_dot_core_dot_Registry__pb2.Registry.FromString,
-                _registered_method=True)
+                )
 
 
 class RegistryServerServicer(object):
@@ -790,7 +770,6 @@ def add_RegistryServerServicer_to_server(servicer, server):
     generic_handler = grpc.method_handlers_generic_handler(
             'feast.registry.RegistryServer', rpc_method_handlers)
     server.add_generic_rpc_handlers((generic_handler,))
-    server.add_registered_method_handlers('feast.registry.RegistryServer', rpc_method_handlers)
 
 
  # This class is part of an EXPERIMENTAL API.
@@ -808,21 +787,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplyEntity',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplyEntity',
             feast_dot_registry_dot_RegistryServer__pb2.ApplyEntityRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetEntity(request,
@@ -835,21 +804,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetEntity',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetEntity',
             feast_dot_registry_dot_RegistryServer__pb2.GetEntityRequest.SerializeToString,
             feast_dot_core_dot_Entity__pb2.Entity.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListEntities(request,
@@ -862,21 +821,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListEntities',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListEntities',
             feast_dot_registry_dot_RegistryServer__pb2.ListEntitiesRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListEntitiesResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def DeleteEntity(request,
@@ -889,21 +838,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/DeleteEntity',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/DeleteEntity',
             feast_dot_registry_dot_RegistryServer__pb2.DeleteEntityRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ApplyDataSource(request,
@@ -916,21 +855,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplyDataSource',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplyDataSource',
             feast_dot_registry_dot_RegistryServer__pb2.ApplyDataSourceRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetDataSource(request,
@@ -943,21 +872,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetDataSource',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetDataSource',
             feast_dot_registry_dot_RegistryServer__pb2.GetDataSourceRequest.SerializeToString,
             feast_dot_core_dot_DataSource__pb2.DataSource.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListDataSources(request,
@@ -970,21 +889,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListDataSources',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListDataSources',
             feast_dot_registry_dot_RegistryServer__pb2.ListDataSourcesRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListDataSourcesResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def DeleteDataSource(request,
@@ -997,21 +906,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/DeleteDataSource',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/DeleteDataSource',
             feast_dot_registry_dot_RegistryServer__pb2.DeleteDataSourceRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ApplyFeatureView(request,
@@ -1024,21 +923,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplyFeatureView',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplyFeatureView',
             feast_dot_registry_dot_RegistryServer__pb2.ApplyFeatureViewRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def DeleteFeatureView(request,
@@ -1051,21 +940,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/DeleteFeatureView',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/DeleteFeatureView',
             feast_dot_registry_dot_RegistryServer__pb2.DeleteFeatureViewRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetAnyFeatureView(request,
@@ -1078,21 +957,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetAnyFeatureView',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetAnyFeatureView',
             feast_dot_registry_dot_RegistryServer__pb2.GetAnyFeatureViewRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.GetAnyFeatureViewResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListAllFeatureViews(request,
@@ -1105,21 +974,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListAllFeatureViews',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListAllFeatureViews',
             feast_dot_registry_dot_RegistryServer__pb2.ListAllFeatureViewsRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListAllFeatureViewsResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetFeatureView(request,
@@ -1132,21 +991,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetFeatureView',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetFeatureView',
             feast_dot_registry_dot_RegistryServer__pb2.GetFeatureViewRequest.SerializeToString,
             feast_dot_core_dot_FeatureView__pb2.FeatureView.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListFeatureViews(request,
@@ -1159,21 +1008,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListFeatureViews',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListFeatureViews',
             feast_dot_registry_dot_RegistryServer__pb2.ListFeatureViewsRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListFeatureViewsResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetStreamFeatureView(request,
@@ -1186,21 +1025,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetStreamFeatureView',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetStreamFeatureView',
             feast_dot_registry_dot_RegistryServer__pb2.GetStreamFeatureViewRequest.SerializeToString,
             feast_dot_core_dot_StreamFeatureView__pb2.StreamFeatureView.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListStreamFeatureViews(request,
@@ -1213,21 +1042,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListStreamFeatureViews',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListStreamFeatureViews',
             feast_dot_registry_dot_RegistryServer__pb2.ListStreamFeatureViewsRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListStreamFeatureViewsResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetOnDemandFeatureView(request,
@@ -1240,21 +1059,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetOnDemandFeatureView',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetOnDemandFeatureView',
             feast_dot_registry_dot_RegistryServer__pb2.GetOnDemandFeatureViewRequest.SerializeToString,
             feast_dot_core_dot_OnDemandFeatureView__pb2.OnDemandFeatureView.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListOnDemandFeatureViews(request,
@@ -1267,21 +1076,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListOnDemandFeatureViews',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListOnDemandFeatureViews',
             feast_dot_registry_dot_RegistryServer__pb2.ListOnDemandFeatureViewsRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListOnDemandFeatureViewsResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ApplyFeatureService(request,
@@ -1294,21 +1093,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplyFeatureService',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplyFeatureService',
             feast_dot_registry_dot_RegistryServer__pb2.ApplyFeatureServiceRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetFeatureService(request,
@@ -1321,21 +1110,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetFeatureService',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetFeatureService',
             feast_dot_registry_dot_RegistryServer__pb2.GetFeatureServiceRequest.SerializeToString,
             feast_dot_core_dot_FeatureService__pb2.FeatureService.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListFeatureServices(request,
@@ -1348,21 +1127,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListFeatureServices',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListFeatureServices',
             feast_dot_registry_dot_RegistryServer__pb2.ListFeatureServicesRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListFeatureServicesResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def DeleteFeatureService(request,
@@ -1375,21 +1144,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/DeleteFeatureService',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/DeleteFeatureService',
             feast_dot_registry_dot_RegistryServer__pb2.DeleteFeatureServiceRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ApplySavedDataset(request,
@@ -1402,21 +1161,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplySavedDataset',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplySavedDataset',
             feast_dot_registry_dot_RegistryServer__pb2.ApplySavedDatasetRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetSavedDataset(request,
@@ -1429,21 +1178,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetSavedDataset',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetSavedDataset',
             feast_dot_registry_dot_RegistryServer__pb2.GetSavedDatasetRequest.SerializeToString,
             feast_dot_core_dot_SavedDataset__pb2.SavedDataset.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListSavedDatasets(request,
@@ -1456,21 +1195,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListSavedDatasets',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListSavedDatasets',
             feast_dot_registry_dot_RegistryServer__pb2.ListSavedDatasetsRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListSavedDatasetsResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def DeleteSavedDataset(request,
@@ -1483,21 +1212,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/DeleteSavedDataset',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/DeleteSavedDataset',
             feast_dot_registry_dot_RegistryServer__pb2.DeleteSavedDatasetRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ApplyValidationReference(request,
@@ -1510,21 +1229,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplyValidationReference',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplyValidationReference',
             feast_dot_registry_dot_RegistryServer__pb2.ApplyValidationReferenceRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetValidationReference(request,
@@ -1537,21 +1246,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetValidationReference',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetValidationReference',
             feast_dot_registry_dot_RegistryServer__pb2.GetValidationReferenceRequest.SerializeToString,
             feast_dot_core_dot_ValidationProfile__pb2.ValidationReference.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListValidationReferences(request,
@@ -1564,21 +1263,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListValidationReferences',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListValidationReferences',
             feast_dot_registry_dot_RegistryServer__pb2.ListValidationReferencesRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListValidationReferencesResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def DeleteValidationReference(request,
@@ -1591,21 +1280,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/DeleteValidationReference',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/DeleteValidationReference',
             feast_dot_registry_dot_RegistryServer__pb2.DeleteValidationReferenceRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ApplyPermission(request,
@@ -1618,21 +1297,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplyPermission',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplyPermission',
             feast_dot_registry_dot_RegistryServer__pb2.ApplyPermissionRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetPermission(request,
@@ -1645,21 +1314,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetPermission',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetPermission',
             feast_dot_registry_dot_RegistryServer__pb2.GetPermissionRequest.SerializeToString,
             feast_dot_core_dot_Permission__pb2.Permission.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListPermissions(request,
@@ -1672,21 +1331,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListPermissions',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListPermissions',
             feast_dot_registry_dot_RegistryServer__pb2.ListPermissionsRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListPermissionsResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def DeletePermission(request,
@@ -1699,21 +1348,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/DeletePermission',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/DeletePermission',
             feast_dot_registry_dot_RegistryServer__pb2.DeletePermissionRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ApplyProject(request,
@@ -1726,21 +1365,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplyProject',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplyProject',
             feast_dot_registry_dot_RegistryServer__pb2.ApplyProjectRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetProject(request,
@@ -1753,21 +1382,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetProject',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetProject',
             feast_dot_registry_dot_RegistryServer__pb2.GetProjectRequest.SerializeToString,
             feast_dot_core_dot_Project__pb2.Project.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListProjects(request,
@@ -1780,21 +1399,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListProjects',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListProjects',
             feast_dot_registry_dot_RegistryServer__pb2.ListProjectsRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListProjectsResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def DeleteProject(request,
@@ -1807,21 +1416,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/DeleteProject',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/DeleteProject',
             feast_dot_registry_dot_RegistryServer__pb2.DeleteProjectRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ApplyMaterialization(request,
@@ -1834,21 +1433,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ApplyMaterialization',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ApplyMaterialization',
             feast_dot_registry_dot_RegistryServer__pb2.ApplyMaterializationRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def ListProjectMetadata(request,
@@ -1861,21 +1450,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/ListProjectMetadata',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/ListProjectMetadata',
             feast_dot_registry_dot_RegistryServer__pb2.ListProjectMetadataRequest.SerializeToString,
             feast_dot_registry_dot_RegistryServer__pb2.ListProjectMetadataResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def UpdateInfra(request,
@@ -1888,21 +1467,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/UpdateInfra',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/UpdateInfra',
             feast_dot_registry_dot_RegistryServer__pb2.UpdateInfraRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetInfra(request,
@@ -1915,21 +1484,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/GetInfra',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/GetInfra',
             feast_dot_registry_dot_RegistryServer__pb2.GetInfraRequest.SerializeToString,
             feast_dot_core_dot_InfraObject__pb2.Infra.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def Commit(request,
@@ -1942,21 +1501,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/Commit',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/Commit',
             google_dot_protobuf_dot_empty__pb2.Empty.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def Refresh(request,
@@ -1969,21 +1518,11 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/Refresh',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/Refresh',
             feast_dot_registry_dot_RegistryServer__pb2.RefreshRequest.SerializeToString,
             google_dot_protobuf_dot_empty__pb2.Empty.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def Proto(request,
@@ -1996,18 +1535,8 @@ class RegistryServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.registry.RegistryServer/Proto',
+        return grpc.experimental.unary_unary(request, target, '/feast.registry.RegistryServer/Proto',
             google_dot_protobuf_dot_empty__pb2.Empty.SerializeToString,
             feast_dot_core_dot_Registry__pb2.Registry.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
