@@ -1,28 +1,8 @@
 # Generated by the gRPC Python protocol compiler plugin. DO NOT EDIT!
 """Client and server classes corresponding to protobuf-defined services."""
 import grpc
-import warnings
 
 from feast.protos.feast.serving import Connector_pb2 as feast_dot_serving_dot_Connector__pb2
-
-GRPC_GENERATED_VERSION = '1.71.0'
-GRPC_VERSION = grpc.__version__
-_version_not_supported = False
-
-try:
-    from grpc._utilities import first_version_is_lower
-    _version_not_supported = first_version_is_lower(GRPC_VERSION, GRPC_GENERATED_VERSION)
-except ImportError:
-    _version_not_supported = True
-
-if _version_not_supported:
-    raise RuntimeError(
-        f'The grpc package installed is at version {GRPC_VERSION},'
-        + f' but the generated code in feast/serving/Connector_pb2_grpc.py depends on'
-        + f' grpcio>={GRPC_GENERATED_VERSION}.'
-        + f' Please upgrade your grpc module to grpcio>={GRPC_GENERATED_VERSION}'
-        + f' or downgrade your generated code using grpcio-tools<={GRPC_VERSION}.'
-    )
 
 
 class OnlineStoreStub(object):
@@ -38,7 +18,7 @@ class OnlineStoreStub(object):
                 '/grpc.connector.OnlineStore/OnlineRead',
                 request_serializer=feast_dot_serving_dot_Connector__pb2.OnlineReadRequest.SerializeToString,
                 response_deserializer=feast_dot_serving_dot_Connector__pb2.OnlineReadResponse.FromString,
-                _registered_method=True)
+                )
 
 
 class OnlineStoreServicer(object):
@@ -62,7 +42,6 @@ def add_OnlineStoreServicer_to_server(servicer, server):
     generic_handler = grpc.method_handlers_generic_handler(
             'grpc.connector.OnlineStore', rpc_method_handlers)
     server.add_generic_rpc_handlers((generic_handler,))
-    server.add_registered_method_handlers('grpc.connector.OnlineStore', rpc_method_handlers)
 
 
  # This class is part of an EXPERIMENTAL API.
@@ -80,18 +59,8 @@ class OnlineStore(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/grpc.connector.OnlineStore/OnlineRead',
+        return grpc.experimental.unary_unary(request, target, '/grpc.connector.OnlineStore/OnlineRead',
             feast_dot_serving_dot_Connector__pb2.OnlineReadRequest.SerializeToString,
             feast_dot_serving_dot_Connector__pb2.OnlineReadResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
