@@ -1,29 +1,9 @@
 # Generated by the gRPC Python protocol compiler plugin. DO NOT EDIT!
 """Client and server classes corresponding to protobuf-defined services."""
 import grpc
-import warnings
 
 from feast.protos.feast.serving import GrpcServer_pb2 as feast_dot_serving_dot_GrpcServer__pb2
 from feast.protos.feast.serving import ServingService_pb2 as feast_dot_serving_dot_ServingService__pb2
-
-GRPC_GENERATED_VERSION = '1.71.0'
-GRPC_VERSION = grpc.__version__
-_version_not_supported = False
-
-try:
-    from grpc._utilities import first_version_is_lower
-    _version_not_supported = first_version_is_lower(GRPC_VERSION, GRPC_GENERATED_VERSION)
-except ImportError:
-    _version_not_supported = True
-
-if _version_not_supported:
-    raise RuntimeError(
-        f'The grpc package installed is at version {GRPC_VERSION},'
-        + f' but the generated code in feast/serving/GrpcServer_pb2_grpc.py depends on'
-        + f' grpcio>={GRPC_GENERATED_VERSION}.'
-        + f' Please upgrade your grpc module to grpcio>={GRPC_GENERATED_VERSION}'
-        + f' or downgrade your generated code using grpcio-tools<={GRPC_VERSION}.'
-    )
 
 
 class GrpcFeatureServerStub(object):
@@ -39,17 +19,17 @@ class GrpcFeatureServerStub(object):
                 '/GrpcFeatureServer/Push',
                 request_serializer=feast_dot_serving_dot_GrpcServer__pb2.PushRequest.SerializeToString,
                 response_deserializer=feast_dot_serving_dot_GrpcServer__pb2.PushResponse.FromString,
-                _registered_method=True)
+                )
         self.WriteToOnlineStore = channel.unary_unary(
                 '/GrpcFeatureServer/WriteToOnlineStore',
                 request_serializer=feast_dot_serving_dot_GrpcServer__pb2.WriteToOnlineStoreRequest.SerializeToString,
                 response_deserializer=feast_dot_serving_dot_GrpcServer__pb2.WriteToOnlineStoreResponse.FromString,
-                _registered_method=True)
+                )
         self.GetOnlineFeatures = channel.unary_unary(
                 '/GrpcFeatureServer/GetOnlineFeatures',
                 request_serializer=feast_dot_serving_dot_ServingService__pb2.GetOnlineFeaturesRequest.SerializeToString,
                 response_deserializer=feast_dot_serving_dot_ServingService__pb2.GetOnlineFeaturesResponse.FromString,
-                _registered_method=True)
+                )
 
 
 class GrpcFeatureServerServicer(object):
@@ -95,7 +75,6 @@ def add_GrpcFeatureServerServicer_to_server(servicer, server):
     generic_handler = grpc.method_handlers_generic_handler(
             'GrpcFeatureServer', rpc_method_handlers)
     server.add_generic_rpc_handlers((generic_handler,))
-    server.add_registered_method_handlers('GrpcFeatureServer', rpc_method_handlers)
 
 
  # This class is part of an EXPERIMENTAL API.
@@ -113,21 +92,11 @@ class GrpcFeatureServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/GrpcFeatureServer/Push',
+        return grpc.experimental.unary_unary(request, target, '/GrpcFeatureServer/Push',
             feast_dot_serving_dot_GrpcServer__pb2.PushRequest.SerializeToString,
             feast_dot_serving_dot_GrpcServer__pb2.PushResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def WriteToOnlineStore(request,
@@ -140,21 +109,11 @@ class GrpcFeatureServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/GrpcFeatureServer/WriteToOnlineStore',
+        return grpc.experimental.unary_unary(request, target, '/GrpcFeatureServer/WriteToOnlineStore',
             feast_dot_serving_dot_GrpcServer__pb2.WriteToOnlineStoreRequest.SerializeToString,
             feast_dot_serving_dot_GrpcServer__pb2.WriteToOnlineStoreResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetOnlineFeatures(request,
@@ -167,18 +126,8 @@ class GrpcFeatureServer(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/GrpcFeatureServer/GetOnlineFeatures',
+        return grpc.experimental.unary_unary(request, target, '/GrpcFeatureServer/GetOnlineFeatures',
             feast_dot_serving_dot_ServingService__pb2.GetOnlineFeaturesRequest.SerializeToString,
             feast_dot_serving_dot_ServingService__pb2.GetOnlineFeaturesResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
