@@ -1,22 +1,12 @@
 # -*- coding: utf-8 -*-
 # Generated by the protocol buffer compiler.  DO NOT EDIT!
-# NO CHECKED-IN PROTOBUF GENCODE
 # source: feast/serving/ServingService.proto
-# Protobuf Python Version: 5.29.0
+# Protobuf Python Version: 4.25.1
 """Generated protocol buffer code."""
 from google.protobuf import descriptor as _descriptor
 from google.protobuf import descriptor_pool as _descriptor_pool
-from google.protobuf import runtime_version as _runtime_version
 from google.protobuf import symbol_database as _symbol_database
 from google.protobuf.internal import builder as _builder
-_runtime_version.ValidateProtobufRuntimeVersion(
-    _runtime_version.Domain.PUBLIC,
-    5,
-    29,
-    0,
-    '',
-    'feast/serving/ServingService.proto'
-)
 # @@protoc_insertion_point(imports)
 
 _sym_db = _symbol_database.Default()
@@ -31,14 +21,14 @@ DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\"feast/serving/Se
 _globals = globals()
 _builder.BuildMessageAndEnumDescriptors(DESCRIPTOR, _globals)
 _builder.BuildTopDescriptorsAndMessages(DESCRIPTOR, 'feast.serving.ServingService_pb2', _globals)
-if not _descriptor._USE_C_DESCRIPTORS:
-  _globals['DESCRIPTOR']._loaded_options = None
+if _descriptor._USE_C_DESCRIPTORS == False:
+  _globals['DESCRIPTOR']._options = None
   _globals['DESCRIPTOR']._serialized_options = b'\n\023feast.proto.servingB\017ServingAPIProtoZ2github.com/feast-dev/feast/go/protos/feast/serving'
-  _globals['_GETONLINEFEATURESREQUESTV2_ENTITYROW_FIELDSENTRY']._loaded_options = None
+  _globals['_GETONLINEFEATURESREQUESTV2_ENTITYROW_FIELDSENTRY']._options = None
   _globals['_GETONLINEFEATURESREQUESTV2_ENTITYROW_FIELDSENTRY']._serialized_options = b'8\001'
-  _globals['_GETONLINEFEATURESREQUEST_ENTITIESENTRY']._loaded_options = None
+  _globals['_GETONLINEFEATURESREQUEST_ENTITIESENTRY']._options = None
   _globals['_GETONLINEFEATURESREQUEST_ENTITIESENTRY']._serialized_options = b'8\001'
-  _globals['_GETONLINEFEATURESREQUEST_REQUESTCONTEXTENTRY']._loaded_options = None
+  _globals['_GETONLINEFEATURESREQUEST_REQUESTCONTEXTENTRY']._options = None
   _globals['_GETONLINEFEATURESREQUEST_REQUESTCONTEXTENTRY']._serialized_options = b'8\001'
   _globals['_FIELDSTATUS']._serialized_start=1560
   _globals['_FIELDSTATUS']._serialized_end=1651
