@@ -1,28 +1,8 @@
 # Generated by the gRPC Python protocol compiler plugin. DO NOT EDIT!
 """Client and server classes corresponding to protobuf-defined services."""
 import grpc
-import warnings
 
 from feast.protos.feast.serving import ServingService_pb2 as feast_dot_serving_dot_ServingService__pb2
-
-GRPC_GENERATED_VERSION = '1.71.0'
-GRPC_VERSION = grpc.__version__
-_version_not_supported = False
-
-try:
-    from grpc._utilities import first_version_is_lower
-    _version_not_supported = first_version_is_lower(GRPC_VERSION, GRPC_GENERATED_VERSION)
-except ImportError:
-    _version_not_supported = True
-
-if _version_not_supported:
-    raise RuntimeError(
-        f'The grpc package installed is at version {GRPC_VERSION},'
-        + f' but the generated code in feast/serving/ServingService_pb2_grpc.py depends on'
-        + f' grpcio>={GRPC_GENERATED_VERSION}.'
-        + f' Please upgrade your grpc module to grpcio>={GRPC_GENERATED_VERSION}'
-        + f' or downgrade your generated code using grpcio-tools<={GRPC_VERSION}.'
-    )
 
 
 class ServingServiceStub(object):
@@ -38,12 +18,12 @@ class ServingServiceStub(object):
                 '/feast.serving.ServingService/GetFeastServingInfo',
                 request_serializer=feast_dot_serving_dot_ServingService__pb2.GetFeastServingInfoRequest.SerializeToString,
                 response_deserializer=feast_dot_serving_dot_ServingService__pb2.GetFeastServingInfoResponse.FromString,
-                _registered_method=True)
+                )
         self.GetOnlineFeatures = channel.unary_unary(
                 '/feast.serving.ServingService/GetOnlineFeatures',
                 request_serializer=feast_dot_serving_dot_ServingService__pb2.GetOnlineFeaturesRequest.SerializeToString,
                 response_deserializer=feast_dot_serving_dot_ServingService__pb2.GetOnlineFeaturesResponse.FromString,
-                _registered_method=True)
+                )
 
 
 class ServingServiceServicer(object):
@@ -80,7 +60,6 @@ def add_ServingServiceServicer_to_server(servicer, server):
     generic_handler = grpc.method_handlers_generic_handler(
             'feast.serving.ServingService', rpc_method_handlers)
     server.add_generic_rpc_handlers((generic_handler,))
-    server.add_registered_method_handlers('feast.serving.ServingService', rpc_method_handlers)
 
 
  # This class is part of an EXPERIMENTAL API.
@@ -98,21 +77,11 @@ class ServingService(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.serving.ServingService/GetFeastServingInfo',
+        return grpc.experimental.unary_unary(request, target, '/feast.serving.ServingService/GetFeastServingInfo',
             feast_dot_serving_dot_ServingService__pb2.GetFeastServingInfoRequest.SerializeToString,
             feast_dot_serving_dot_ServingService__pb2.GetFeastServingInfoResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def GetOnlineFeatures(request,
@@ -125,18 +94,8 @@ class ServingService(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.serving.ServingService/GetOnlineFeatures',
+        return grpc.experimental.unary_unary(request, target, '/feast.serving.ServingService/GetOnlineFeatures',
             feast_dot_serving_dot_ServingService__pb2.GetOnlineFeaturesRequest.SerializeToString,
             feast_dot_serving_dot_ServingService__pb2.GetOnlineFeaturesResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
