@@ -1,28 +1,8 @@
 # Generated by the gRPC Python protocol compiler plugin. DO NOT EDIT!
 """Client and server classes corresponding to protobuf-defined services."""
 import grpc
-import warnings
 
 from feast.protos.feast.serving import TransformationService_pb2 as feast_dot_serving_dot_TransformationService__pb2
-
-GRPC_GENERATED_VERSION = '1.71.0'
-GRPC_VERSION = grpc.__version__
-_version_not_supported = False
-
-try:
-    from grpc._utilities import first_version_is_lower
-    _version_not_supported = first_version_is_lower(GRPC_VERSION, GRPC_GENERATED_VERSION)
-except ImportError:
-    _version_not_supported = True
-
-if _version_not_supported:
-    raise RuntimeError(
-        f'The grpc package installed is at version {GRPC_VERSION},'
-        + f' but the generated code in feast/serving/TransformationService_pb2_grpc.py depends on'
-        + f' grpcio>={GRPC_GENERATED_VERSION}.'
-        + f' Please upgrade your grpc module to grpcio>={GRPC_GENERATED_VERSION}'
-        + f' or downgrade your generated code using grpcio-tools<={GRPC_VERSION}.'
-    )
 
 
 class TransformationServiceStub(object):
@@ -38,12 +18,12 @@ class TransformationServiceStub(object):
                 '/feast.serving.TransformationService/GetTransformationServiceInfo',
                 request_serializer=feast_dot_serving_dot_TransformationService__pb2.GetTransformationServiceInfoRequest.SerializeToString,
                 response_deserializer=feast_dot_serving_dot_TransformationService__pb2.GetTransformationServiceInfoResponse.FromString,
-                _registered_method=True)
+                )
         self.TransformFeatures = channel.unary_unary(
                 '/feast.serving.TransformationService/TransformFeatures',
                 request_serializer=feast_dot_serving_dot_TransformationService__pb2.TransformFeaturesRequest.SerializeToString,
                 response_deserializer=feast_dot_serving_dot_TransformationService__pb2.TransformFeaturesResponse.FromString,
-                _registered_method=True)
+                )
 
 
 class TransformationServiceServicer(object):
@@ -78,7 +58,6 @@ def add_TransformationServiceServicer_to_server(servicer, server):
     generic_handler = grpc.method_handlers_generic_handler(
             'feast.serving.TransformationService', rpc_method_handlers)
     server.add_generic_rpc_handlers((generic_handler,))
-    server.add_registered_method_handlers('feast.serving.TransformationService', rpc_method_handlers)
 
 
  # This class is part of an EXPERIMENTAL API.
@@ -96,21 +75,11 @@ class TransformationService(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.serving.TransformationService/GetTransformationServiceInfo',
+        return grpc.experimental.unary_unary(request, target, '/feast.serving.TransformationService/GetTransformationServiceInfo',
             feast_dot_serving_dot_TransformationService__pb2.GetTransformationServiceInfoRequest.SerializeToString,
             feast_dot_serving_dot_TransformationService__pb2.GetTransformationServiceInfoResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
 
     @staticmethod
     def TransformFeatures(request,
@@ -123,18 +92,8 @@ class TransformationService(object):
             wait_for_ready=None,
             timeout=None,
             metadata=None):
-        return grpc.experimental.unary_unary(
-            request,
-            target,
-            '/feast.serving.TransformationService/TransformFeatures',
+        return grpc.experimental.unary_unary(request, target, '/feast.serving.TransformationService/TransformFeatures',
             feast_dot_serving_dot_TransformationService__pb2.TransformFeaturesRequest.SerializeToString,
             feast_dot_serving_dot_TransformationService__pb2.TransformFeaturesResponse.FromString,
-            options,
-            channel_credentials,
-            insecure,
-            call_credentials,
-            compression,
-            wait_for_ready,
-            timeout,
-            metadata,
-            _registered_method=True)
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
